Bug fixed!

+++ Exercises/Polymorphism - Bank account question/src/SavingsAccount.java
@@ -3,14 +3,15 @@ public class SavingsAccount extends BankAccount {
         super(accountNumber, balance, usrPassword);
     }
 
-    //override function
-    public void deposit(int amount, String usrPassword) {
+    // override function
+    // note to self: if override function's parameters should be SAME!
+    public void deposit(double amount, String usrPassword) {
         if (usrPassword == this.usrPassword) {
             double interest = (this.balance / 100) * 3;
-            this.balance += amount + interest ;
+            this.balance += amount + interest;
 
             System.out.println("$" + amount + " deposited, $" + interest + " interest added!");
-            
+
             System.out.println("New balance is = $" + balance);
             System.out.println();
         } else {
